FITFRND-23 UserIdResolver 로직 변경
- user-service 패키지, 클래스 명 정리
