+++ back/user-service/src/main/java/com/example/userservice/common/auth/SocialOAuth.java
@@ -1,9 +1,5 @@
 package com.example.userservice.common.auth;
 
 public interface SocialOAuth {
-    String getOAuthRedirectUrl();
-
-    String requestAccessToken(String code);
-
     String getUserInfo(String code);
 }
