+++ back/user-service/src/main/java/com/example/userservice/common/resolver/UserIdResolver.java
@@ -1,10 +1,8 @@
 package com.example.userservice.common.resolver;
 
 
-import com.example.userservice.common.auth.JwtAuthProvider;
-import com.example.userservice.common.resolver.memberid.MemberId;
+import com.example.userservice.common.resolver.userid.UserId;
 import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.stereotype.Component;
@@ -14,30 +12,22 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 @Component
-public class MemberIdResolver implements HandlerMethodArgumentResolver {
-
-    @Autowired
-    private JwtAuthProvider jwtAuthProvider;
+public class UserIdResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        MemberId annotation = parameter.getParameterAnnotation(MemberId.class);
-
-
-
+        UserId annotation = parameter.getParameterAnnotation(UserId.class);
         return annotation != null && parameter.getParameterType().equals(Long.class);
     }
 
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
+        String authorization = request.getHeader("User-Id");
 
-        String authorization = request.getHeader("Authorization");
-
-        if(!jwtAuthProvider.validateToken(authorization)) {
+        if(authorization.isEmpty()) {
             throw new AuthorizationDeniedException("Invalid token");
         }
-
-        return jwtAuthProvider.getIdByToken(authorization);
+        return authorization;
     }
 }
