+++ back/user-service/src/main/java/com/example/userservice/common/resolver/userid/UserId.java
@@ -1,4 +1,4 @@
-package com.example.userservice.common.resolver.memberid;
+package com.example.userservice.common.resolver.userid;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.PARAMETER)
-public @interface MemberId {
+public @interface UserId {
 }
