+++ back/user-service/src/main/java/com/example/userservice/controller/UserController.java
@@ -1,16 +1,17 @@
 package com.example.userservice.controller;
 
 import com.example.userservice.common.dto.CustomResponseBody;
-import com.example.userservice.common.resolver.memberid.MemberId;
+import com.example.userservice.common.resolver.userid.UserId;
 import com.example.userservice.common.util.ResponseUtil;
+import com.example.userservice.domain.User;
 import com.example.userservice.dto.LoadUserDetailResponse;
 import com.example.userservice.dto.ModifyUserDetailRequest;
-import com.example.userservice.service.UserDetailService;
+import com.example.userservice.dto.SaveUserRequest;
+import com.example.userservice.service.UserService;
 import com.fasterxml.jackson.databind.ser.FilterProvider;
 import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.json.MappingJacksonValue;
 import org.springframework.web.bind.annotation.*;
@@ -19,20 +20,29 @@ import java.util.UUID;
 
 
 @RestController
+@RequestMapping("/")
+@RequiredArgsConstructor
 @CrossOrigin(origins="http://localhost:19006", allowedHeaders = "*")
 public class UserController {
+    private final UserService userService;
 
-    private final UserDetailService userDetailService;
 
-    @Autowired
-    public UserController(UserDetailService userDetailService) {
-        this.userDetailService = userDetailService;
+    @PostMapping("/signup")
+    public ResponseEntity<User> signUp(@RequestBody SaveUserRequest request) {
+        User save = userService.save(request);
+        try {
+            return ResponseEntity.ok()
+                    .body(save);
+        }catch (Exception e){
+            return ResponseEntity.internalServerError()
+                    .body(save);
+        }
     }
 
-    @GetMapping("/api/user/{userId}")
-    ResponseEntity<CustomResponseBody<MappingJacksonValue>> loadUserDetail(@PathVariable UUID userId, @MemberId UUID memberId) {
+    @GetMapping("/user/{userId}")
+    ResponseEntity<CustomResponseBody<MappingJacksonValue>> loadUserDetail(@PathVariable UUID userId, @RequestHeader("User-Id") String memberId) {
 
-        LoadUserDetailResponse user = userDetailService.findUser(userId, memberId);
+        LoadUserDetailResponse user = userService.findUser(userId, UUID.fromString(memberId));
         SimpleBeanPropertyFilter filter = SimpleBeanPropertyFilter.filterOutAllExcept("matchId","category","headCnt","place","tag","startTime","endTime","attendanceCnt");
         FilterProvider filters = new SimpleFilterProvider().addFilter("UserInfo", filter);
 
@@ -42,9 +52,9 @@ public class UserController {
         return ResponseUtil.success(mapping);
     }
 
-    @PutMapping("/api/user")
-    String modifyUserDetail(@MemberId UUID memberId ,@RequestBody ModifyUserDetailRequest request) {
-        return userDetailService.modifyUserDetail(request, memberId);
+    @PutMapping("/user")
+    String modifyUserDetail(@UserId UUID memberId , @RequestBody ModifyUserDetailRequest request) {
+        return userService.modifyUserDetail(request, memberId);
 
     }
 }
