+++ back/user-service/src/main/java/com/example/userservice/dto/UserIdResponse.java
@@ -1,0 +1,10 @@
+package com.example.userservice.dto;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+public class UserIdResponse {
+    private UUID userId;
+}
