+++ back/user-service/src/main/resources/application.yml
@@ -26,14 +26,6 @@ eureka:
     service-url:
       defaultZone: ${EUREKA.SERVICE.URL}
 google:
-  client:
-    id: ${GOOGLE_CLIENTID}
-    secret: ${GOOGLE_CLIENT_SECRET}
-  redirect:
-    uri: ${GOOGLE_REDIRECT_URI}
-  base:
-    uri: ${GOOGLE_BASE_URI}
-  token_uri: ${GOOGLE_TOKEN_URI}
   userInfo_uri: ${GOOGLE_USERINFO_URI}
 logging:
   level:
